+++ tools/daemon/vars.do.pl
@@ -6,6 +6,9 @@
 # @(-) --[no]verbose           run verbosely [${verbose}]
 # @(-) --[no]confDirs          display the list of directories which may contain daemons configurations [${confDirs}]
 # @(-) --[no]execDirs          display the list of directories which may contain daemons executables [${execDirs}]
+# @(-) --json=<name>           the JSON file to operate on when requesting by keys [${json}]
+# @(-) --name=<name>           the daemon name to operate on when requesting by keys [${name}]
+# @(-) --key=<name[,...]>      the key which addresses the desired value, may be specified several times or as a comma-separated list [${key}]
 #
 # TheToolsProject - Tools System and Working Paradigm for IT Production
 # Copyright (©) 1998-2023 Pierre Wieser (see AUTHORS)
@@ -30,6 +33,7 @@ use utf8;
 use warnings;
 
 use TTP::DaemonConfig;
+use TTP::Finder;
 
 my $defaults = {
 	help => 'no',
@@ -37,11 +41,28 @@ my $defaults = {
 	dummy => 'no',
 	verbose => 'no',
 	confDirs => 'no',
-	execDirs => 'no'
+	execDirs => 'no',
+	json => '',
+	name => '',
+	key => ''
 };
 
 my $opt_confDirs = false;
 my $opt_execDirs = false;
+my $opt_json = $defaults->{json};
+my $opt_name = $defaults->{name};
+my @opt_keys = ();
+
+# the daemon config when requesting by keys
+my $daemonConfig = undef;
+
+# -------------------------------------------------------------------------------------------------
+# Display the value accessible through the route of the provided successive keys
+
+sub listByKeys {
+	my $value = $daemonConfig->var( \@opt_keys );
+	print "  [".join( ',', @opt_keys )."]: ".( defined( $value ) ? ( ref( $value ) ? Dumper( $value ) : $value.EOL ) : "(undef)".EOL );
+}
 
 # -------------------------------------------------------------------------------------------------
 # list confDirs value - e.g. 'C:\INLINGUA\configurations\daemons'
@@ -87,7 +108,10 @@ if( !GetOptions(
 	"dummy!"			=> sub { $ep->runner()->dummy( @_ ); },
 	"verbose!"			=> sub { $ep->runner()->verbose( @_ ); },
 	"confDirs!"			=> \$opt_confDirs,
-	"execDirs!"			=> \$opt_execDirs )){
+	"execDirs!"			=> \$opt_execDirs,
+	"json=s"			=> \$opt_json,
+	"name=s"			=> \$opt_name,
+	"key=s"				=> \@opt_keys )){
 
 		msgOut( "try '".$ep->runner()->command()." ".$ep->runner()->verb()." --help' to get full usage syntax" );
 		TTP::exit( 1 );
@@ -103,10 +127,44 @@ msgVerbose( "got dummy='".( $ep->runner()->dummy() ? 'true':'false' )."'" );
 msgVerbose( "got verbose='".( $ep->runner()->verbose() ? 'true':'false' )."'" );
 msgVerbose( "got confDirs='".( $opt_confDirs ? 'true':'false' )."'" );
 msgVerbose( "got execDirs='".( $opt_execDirs ? 'true':'false' )."'" );
+msgVerbose( "got json='$opt_json'" );
+msgVerbose( "got name='$opt_name'" );
+@opt_keys= split( /,/, join( ',', @opt_keys ));
+msgVerbose( "got keys='".join( ',', @opt_keys )."'" );
+
+# when requesting keys, they must be addressed inside of a daemon configuration, so have a json or a name
+if( scalar( @opt_keys )){
+	my $count = 0;
+	$count += 1 if $opt_json;
+	$count += 1 if $opt_name;
+	if( $count == 0 ){
+		msgErr( "one of '--json' or '--name' options must be specified when requesting by keys, none found" );
+	} elsif( $count > 1 ){
+		msgErr( "one of '--json' or '--name' options must be specified when requesting by keys, both were found" );
+	}
+	# if a daemon name is specified, find the full filename of the JSON configuration file
+	if( $opt_name ){
+		my $finder = TTP::Finder->new( $ep );
+		my $confFinder = TTP::DaemonConfig->confFinder();
+		$opt_json = $finder->find({ dirs => [ $confFinder->{dirs}, $opt_name ], suffix => $confFinder->{suffix}, wantsAll => false });
+		msgErr( "unable to find a suitable daemon JSON configuration file for '$opt_name'" ) if !$opt_json;
+	}
+	# if a json has been specified or has been found, must be loadable
+	if( $opt_json ){
+		$daemonConfig = TTP::DaemonConfig->new( $ep, { jsonPath => $opt_json, checkConfig => false });
+		if( !$daemonConfig->jsonLoaded()){
+			msgErr( "unable to load a suitable daemon configuration for json='$opt_json'" );
+		}
+	}
+}
+
+# warn if no option has been requested
+msgWarn( "none of '--confDirs', '--execDirs', '--key' options has been requested, nothing to do" ) if !$opt_confDirs && !$opt_execDirs && !scalar( @opt_keys );
 
 if( !TTP::errs()){
 	listConfdirs() if $opt_confDirs;
 	listExecdirs() if $opt_execDirs;
+	listByKeys() if scalar @opt_keys;
 }
 
 TTP::exit();